requirements.txt
tensorflow-cpu==2.10
tensorflow-addons
opencv-python-headless
scikit-image
scipy
gradio
